+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Any non-code changes should be prefixed with `(docs)`.
 See `PUBLISH.md` for instructions on how to publish a new version.
 -->
 
+- (minor) Add Limit tokens plugin
 - (minor) Add Image Compare embeds
 - (minor) Add Slideshow embeds
 - (minor) Add Instagram embeds